Human AI with bugs
