+++ Assets/Scripts/HumanBT-AI/Conquer.cs
@@ -6,7 +6,6 @@ using BehaviorTree;
 
 public class Conquer : Node 
 {
-    PathFinding pathfindingClass = new PathFinding();
 
     // Only engage in conquering if you're outnumbered (equivalent of retreating)
     public override NodeState Evaluate()
@@ -52,7 +51,7 @@ public class Conquer : Node
                 }
 
                 // Find a path (if no valid path just continue and stay put)
-                List<GridNode> path = pathfindingClass.A_Star(ref tempHum, destination);
+                List<GridNode> path = A_Star(ref tempHum, destination);
                 if(path == null)
                     continue;
 
@@ -78,7 +77,7 @@ public class Conquer : Node
                 }
 
                 // Find a path (if no valid path just continue and stay put)
-                List<GridNode> path = pathfindingClass.A_Star(ref tempHum, destination);
+                List<GridNode> path = A_Star(ref tempHum, destination);
                 if(path == null)
                     continue;
 
@@ -91,13 +90,6 @@ public class Conquer : Node
 
     }
 
-    float GetDistance(int[] coordsA, int[] coordsB) {
-        float dstX = Mathf.Pow((coordsB[0] - coordsA[0]), 2);
-        float dstY = Mathf.Pow((coordsB[1] - coordsA[1]), 2);
-
-        return Mathf.Sqrt(dstX + dstY);
-    }
-
     void Move(List<GridNode> path, ref Human agentToMove) {
         // We need to start by ensuring the grid we are standing on no longer references us in any way
         int[] coords = agentToMove.Coordinates;
@@ -105,16 +97,16 @@ public class Conquer : Node
         GameManager.Instance.CoordsToGridNode[(coords[0], coords[1])].Occupation = 0;
         GameManager.Instance.CoordsToGridNode[(coords[0], coords[1])].Standing = null;
 
+        float walkCooldown = 2.0f;
+        float currWalkCooldown = 0.0f;
+
         // Move from grid to grid
         foreach(GridNode node in path) {
-            Vector3 targetPosition = new Vector3(node.gameObject.transform.localPosition.x, 0, node.gameObject.transform.localPosition.z);
-            Vector3 velocity = Vector3.zero;
-            agentToMove.gameObject.transform.localPosition = Vector3.SmoothDamp(
-                agentToMove.gameObject.transform.localPosition,
-                targetPosition,
-                ref velocity,
-                0.3f
-            );
+            Vector3 targetPosition = new Vector3(node.gameObject.transform.localPosition.x, agentToMove.gameObject.transform.localPosition.y, node.gameObject.transform.localPosition.z);
+            agentToMove.gameObject.transform.localPosition = targetPosition;
+            while(currWalkCooldown < walkCooldown)
+                currWalkCooldown += Time.deltaTime;
+            currWalkCooldown = 0.0f;
         }
 
         // Once we are on last grid we need to update the grid we are standing on, our coords, everything
@@ -128,6 +120,135 @@ public class Conquer : Node
         GameManager.Instance.CoordsToGridNode[(copyNodeCoords[0], copyNodeCoords[1])].Standing = agentToMove.gameObject;
 
         return;
+    }
+
+    
+    public class PathNode {
+        public GridNode _Node;
+        public int gCost;
+        public int hCost;
+        public PathNode parent;
+
+        public PathNode(GridNode node) {
+            this._Node = node;
+            this.gCost = 1000000;
+            this.hCost = 1000000;
+            this.parent = null;
+        }
+
+        public int fCost {
+            get {
+                return this.gCost + this.hCost;
+            }
+        }
+
+
+    }
+
+    // Returns full path, it's up to the constraints of the individual to just go the amount of steps they CAN go to
+    public List<GridNode> A_Star(ref Human agentToMove, GridNode destination) {
+        int[] coords = agentToMove.Coordinates;
+
+        List<PathNode> openSet = new List<PathNode>();
+        HashSet<PathNode> closedSet = new HashSet<PathNode>();
+
+        PathNode startNode = new PathNode(GameManager.Instance.CoordsToGridNode[(coords[0], coords[1])]);
+        PathNode destNode = new PathNode(destination);
+        //Debug.Log(startNode._Node.Coordinates[0] + " " + startNode._Node.Coordinates[1]);
+        //Debug.Log(destNode._Node.Coordinates[0] + " " + destNode._Node.Coordinates[1]);
+        startNode.gCost = 0;
+        startNode.hCost = (int)GetDistance(startNode, destNode);
+        destNode.hCost = 0;
+        openSet.Add(startNode);
+
+        while (openSet.Count > 0) {
+            PathNode curr = openSet[0];
+            for (int i = 1; i < openSet.Count; i++) {
+                if (openSet[i].fCost < curr.fCost || (openSet[i].fCost == curr.fCost && openSet[i].hCost < curr.hCost)) {
+                    curr = openSet[i];
+                }
+            }
+
+            openSet.Remove(curr);
+            closedSet.Add(curr);
+
+            if(curr._Node.Coordinates[0] == destNode._Node.Coordinates[0] && curr._Node.Coordinates[1] == destNode._Node.Coordinates[1]) {
+                return RetracePath(startNode, destNode);
+            }
+
+            
+            List<PathNode> neighbours = GetNeighbors(curr);
+
+            foreach(PathNode neighbour in neighbours) {
+                if(neighbour._Node.Occupation != 0 || closedSet.Contains(neighbour)) {
+                    continue;
+                }
+
+                int newMovementCostToNeighbour = curr.gCost + (int)GetDistance(curr, neighbour);
+                if(newMovementCostToNeighbour < neighbour.gCost || !openSet.Contains(neighbour)) {
+                    neighbour.gCost = newMovementCostToNeighbour;
+                    neighbour.hCost = (int)GetDistance(neighbour, destNode);
+                    neighbour.parent = curr;
+
+
+                    if(!openSet.Contains(neighbour))
+                        openSet.Add(neighbour);
+                }
+
+
+            }
+        }
+
+        return null;
+    }
+
+    public List<PathNode> GetNeighbors(PathNode curr) {
+        List<PathNode> neighbours = new List<PathNode>();
+        for(int i = -1; i <= 1; i++){
+            for(int j = -1; j <= 1; j++) {
+                if(i ==0 && j == 0)
+                    continue;
+
+                int[] coords = curr._Node.Coordinates;
+                if (!(GameManager.Instance.CoordsToGridNode.ContainsKey((coords[0] + i, coords[1] + j))))
+                    continue;
+                else {
+                    PathNode newPathNode = new PathNode(GameManager.Instance.CoordsToGridNode[(coords[0] + i, coords[1] + j)]);
+                    neighbours.Add(newPathNode);
+                }
+            }
+        }
+
+        return neighbours;
+    }
+
+    List<GridNode> RetracePath(PathNode startNode, PathNode endNode) {
+        List<GridNode> path = new List<GridNode>();
+        PathNode currNode = endNode;
+        GridNode currGridNode = currNode._Node;
+        GridNode startGridNode = startNode._Node;
+
+        while (currNode != null && currGridNode != startGridNode) {
+            path.Add(currGridNode);
+            currNode = currNode.parent;
+        }
+
+        path.Reverse();
+        return path;
+    }
+
+    float GetDistance(PathNode nodeA, PathNode nodeB) {
+        float dstX = Mathf.Pow((nodeB._Node.Coordinates[0] - nodeA._Node.Coordinates[0]), 2);
+        float dstY = Mathf.Pow((nodeB._Node.Coordinates[1] - nodeA._Node.Coordinates[1]), 2);
+
+        return Mathf.Sqrt(dstX + dstY);
+    }
+
+    float GetDistance(int[] coordsA, int[] coordsB) {
+        float dstX = Mathf.Pow((coordsB[0] - coordsA[0]), 2);
+        float dstY = Mathf.Pow((coordsB[1] - coordsA[1]), 2);
+
+        return Mathf.Sqrt(dstX + dstY);
     }
 
 }
