--- Assets/Scripts/HumanBT-AI/PathFinding.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a744ecc9d8e31a94bba9a67cf83490a4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
