+++ Assets/Scripts/SpawnManager.cs
@@ -53,6 +53,7 @@ public class SpawnManager : MonoBehaviour
     
         }
         else if (spawnSpecification == "Hum"){
+            Debug.Log("kys");
             temp = GameObject.Instantiate(_HumSpawn, position, transform.rotation);
             
             Human tempHum = temp.GetComponent(typeof(Human)) as Human;
